refactor: decompose admin panel workflows
